requirements.txt
numpy==1.26.4
streamlit==1.32.0
tensorflow-cpu==2.16.2
keras
pillow